+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Added
 
+- added agriculture class to land use
 - split basic and detailed reports into two different functions
 - added sort to detailed reports to ensure rows in correct order regardless of osm data
 
 ### Changed
 
+- relabeled unknown class to other
 - combined unpaved and paved road types into one class
 
 ### Added